refactor(admin): separate admin tabs
